pip_requirements.txt
krpc==0.5.4
protobuf==6.31.1
pyserial==3.5
